Secure signature-policy default and options
Set the default signature-policy to require-signatures because software update mechanisms need to be secure.

Remove the verify option because it provides zero security, can't implement policy, and the name is incredibly misleading. Any attacker who can modify the repo, either by MITM or access to the repo server, just needs to remove the signature from the manifest to be able alter a package.

Users currently using 'verify' should be honest about it and use 'ignore'.